requirements.txt
requests==2.31.0
fastapi==0.95.1
uvicorn==0.22.0
mongoengine==0.27.0
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7
bcrypt==4.0.1
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
logging==0.4.9.6
flake8==7.1.2
pytz==2025.1